Move wakeonlan.py to __init__.py
